backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart>=0.0.6
asyncpg>=0.29.0
Pillow>=10.0.0
numpy>=1.24.0
anthropic>=0.18.0
APScheduler>=3.10.0
